Define a top menu using hugo's menus
We are not using main menu because it is quite coupled with sections but
we want to link to pages internal and external pages directly, for
example to home, github, etc.

This also implments the required Go template to mark the active menu
entry which we do by adding the active class.