Add primary font
We use hugo resources to embed the css file containing the google
generated @font-family and also to embed the fonts as base64 directly
into that css file.

The result is 0 new requests are needed and font will be applied almost
instantly so we do not have the headache of having to decide how to
handle font-display to avoid flash of invisible text (FOIT) or similar
jumpy effects.