Use helaili/jekyll-action@v2 for jekyll
To make the environment 100% reproducible and without duplication we
were trying to use the same Makefile we use in development but github is
being stubborn with permission errors.

Since all these changes are preciselly about removing Jekyll from the
equation it does not make sense to invest more time debugging this.

elaili/jekyll-action@v2 code seems simple enough we could even fork it
if required.