Unify top navigation into the same flex-box
We want to implement a Priority+ navigation pattern using only CSS so we
need the browser to help us out by overflowing things automatically when
required.

Before this commit the navigation was split in two elements containing
each a list of links which made it impossible for the browser to do the
right thing of hiding/wraping just one element when no enough space as
available.

Also, we are now using a list to aid in accessibility and to make the
site somewhat useful when no css is available.