Unify icons by using tabler icon theme
We are adding a mode where the nav-bar will show icons and text for all
the entries and we are using the opportunity not to only add the newly
required icons but also to unify them.

We are using the awesome https://tabler-icons.io/ and modify some of
them to fit our style.