First attempt into creating a Priority+ navigation only with CSS
The code might be bit messy, we will do the refactor once the menu looks
exactly like we want it to.

What this commit does is it adds a hidden checkbox and a bunch of styles
that will only activate when the :checked selector is enabled.

This commit makes the navigation responsive enough to continue with the
rest of the theme althoug its UX is far from perfect.