Use fluid typo
So while researching what would be a good clamp formula for font size we
stumbled on https://www.fluid-type-scale.com/ which generates the
formula according to the given specification.

It basically has a little bit of the viewport with into account so the
fonts can get bigger/smaller depending on it and then uses clamp so
things do not get ridiculously small or large.

The default seems to work great for us.