Add aliases to match old urls
The blog is linked in some places and it wouldn't be nice if we break
the url's.

It also adds a slug generated from the filename as it used to be the
case in jekyll.

For now we added the prefix articles/ to the permalinks but that is just
a placeholder.