Promote hugo+plainlyColorful as the main site
The new site is still a huge WIP but I already like it way better than
the old one.

In theory Urls compatibility is in place.
RSS is NOT compatible but that has low prio since KDE planet is no
longuer indexing the site.

Most main navigation menu is not working either.

I'd say, ship it! 🥳