Use flex gap instead of margin in the nav-top-primary menu
Flex gap is also applied for columns when the wrap kicks in which means
that we can be sure that the contents of the nav elements will not
overlap when the wrap happens.

Ultimately this makes the nav items height to be irrelevant, opening the
door to add more stuff in there like a decorator that shows which tab is
active.