Merge theme and blog
When we started to migrate to hugo we developed the templating/design in
a theme because we did not know better. Also, it was useful to easily
test things.

Now time has passed and we do not have any reason to keep the theme
around since we do not have any plan of reusing it. Further more, if we
ever want to have it again we can just undo this commit since the
separation of content and layout in Hugo makes it really really
difficult to couple both.