Use file enumeration again to generate icons classes
It made sense to use menus when there was only one use of each icon and
when all the icons were use.

But now we are already re-using icons in different menus plus at least
one icon (color toggle) might not be in any menu but we still want the
icon to be embeded.