Fix link animations in description
The generic A selector was selecting the nav items and performing two
animations which did not match.

For now we are applying a hack by using :not(.nav). We will figure out
how to do it properly later.

Also, We are now scaling the links themselves by using
display:inline-block instead of their container.