Improve LQIP util classes
The improvement though is made by forcing us to use a wrapper object
since our current solution was not working on FF and probably fell out
of the standard.

This commit uses the initial approach of having a container with an
::affter pseudo element positioned absolutly with top and left 0;

This time around though we are using a Picture element so things are a
bit more semantic.