Fix navar items vertical alignment
We were focing all .nav children to be inline-box which worked just fine
for text/inline elements but it was making the icons larger since the
line-height was being taken into account.

Also, we had a mix bag of layoutingin the navvar which was not helpiong
while debugging this. This ocmmit also simplifies it.