Add dynamic content width
We want each part of our content to define their own width as long as it
is constrained within our --max-content-layout limits.

This commit removes the hard constrain and instead it replaces it with
--content-width which is just the defualt now but every element inside
it can override it.

Also it removes the content div which basically was the element
enforcing the constraint.