Adjust max-width depending on aspect-ratio
When the viewport has enough room to accomodate wider images we should
make use of that extra space to display narrower images in a way that is
more confortable.

The way this it is implemented basically makes 1:1 images just use the
--content-width as max-width while anything greater (1:1.1) will use
more horizontal space until --max-content-layout is reached.

This makes horizontal images more confortable and ergonomic since they
will make use of the extra space.

This is not implemented yet for vertical images. We'll do that when we
need it.