Make new image rendering opt-in and make it work on mobile
If an image has a -lqip.webp version then the system kick's in,
otherwise it assumes that the passed image does not support any of the
fancy features.

This commit contains more changes than intended, it basically checks for
.avif files as well.

BTW hugo does not have .avif file support so we are using .webp for
extracting the size and then just adding .avif with the same
information.