Underline article list
This required way more work than expected since we have a bit of a mess
with generic selectors that will pick up almost any link.

For now what we've done is to add a new utility class called .scale
which takes care of the scaling, then that class is only used by header.

Also in order to reuse "nav home" we've had to add a span to the links
in the index.