Use nav only for Site menus
It does not matter if the link starts with .BaseUrl, we won't add nav
unless the link is in the menu.

This fixes a bit the mess that we've got with nav and a.

Although the proper fix is probably stop messing around in the markdown
render hook and use actual html for the about section.