Publish an old Kafka post from 2018
It got lost somewhere in the migration between
wordpress to jekyll and it was never published.

Its content is outdated but imho it still is
a bit relevant since the initial kafka trade-off
cascaded many of the solutions that we can use
today like Kafka connect, KSQL or streams.

This is the first of 10 articles that I will be
publishing after the fact becuase either they
were kept as drafts, written for internal use
or else.